[UPSTREAM MERGE] merge back esr128:b828740 into Firefox. It's now based off Firefox Stable! closes rustdevbtw/flamewolf#1

+++ media/libvpx/libvpx/vp8/encoder/onyx_int.h
@@ -623,7 +623,7 @@ typedef struct VP8_COMP {
   double totalp_v;
   double totalp;
   double total_sq_error2;
-  uint64_t bytes;
+  int bytes;
   double summed_quality;
   double summed_weights;
   unsigned int tot_recode_hits;
